requirements.txt
altgraph==0.17.4
aniso8601==9.0.1
arrow==1.3.0
Babel==2.13.1
blinker==1.7.0
build==1.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
Flask==3.0.0
flask-babel==4.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-MQTT==1.1.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
paho-mqtt==1.6.1
pandas==2.1.3
passlib==1.7.4
pefile==2023.2.7
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
tomli==2.0.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
waitress==2.1.2
Werkzeug==3.0.1
zipp==3.17.0
